fix: signing installer and .exe
